+++ src/ExplainableAI.jl
@@ -11,8 +11,7 @@ using ProgressMeter: Progress, next!
 
 # Automatic differentiation
 using ADTypes: AbstractADType, AutoZygote
-using DifferentiationInterface: value_and_gradient
-using DifferentiationInterface: prepare_pullback, value_and_pullback, value_and_pullback!
+import DifferentiationInterface as DI
 const DEFAULT_AD_BACKEND = AutoZygote()
 
 include("bibliography.jl")